backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
Pillow>=10.4.0
pytesseract>=0.3.13
pydantic>=2.8.0
python-dotenv>=1.0.1
python-docx>=1.1.2
supabase>=2.5.1
aiohttp>=3.9.5
rapidfuzz>=3.9.0
textstat>=0.7.3
python-dateutil>=2.9.0
